:rocket: refactored some dto's, services and contr

+++ src/main/java/br/com/serratec/serramed/domain/service/EnderecoService.java
@@ -1,8 +1,8 @@
 package br.com.serratec.serramed.domain.service;
 
+import br.com.serratec.serramed.domain.exception.BadRequestException;
 import br.com.serratec.serramed.domain.exception.NotFoundException;
 import br.com.serratec.serramed.domain.model.Endereco;
-import br.com.serratec.serramed.domain.model.Hospital;
 import br.com.serratec.serramed.domain.repository.EnderecoRepository;
 import br.com.serratec.serramed.domain.service.iCrud.ICRUDService;
 import br.com.serratec.serramed.dto.endereco.EnderecoRequestDto;
@@ -13,6 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -23,26 +24,20 @@ public class EnderecoService implements ICRUDService<EnderecoRequestDto, Enderec
 	private EnderecoRepository enderecoRepository;
 
 	@Autowired
-	private HospitalService hospitalService;
-
-	@Autowired
 	private ModelMapper mapper;
 
 	@Override
 	public EnderecoResponseDto create(EnderecoRequestDto dto) {
-		Hospital hospital = hospitalService.findById(dto.getHospitalId());
+		
 		Endereco endereco = mapper.map(dto, Endereco.class);
-		endereco.setHospital(hospital);
 		return mapper.map(enderecoRepository.save(endereco), EnderecoResponseDto.class);
 	}
 
 	@Override
 	public EnderecoResponseDto findById(Long id) {
-		Optional<Endereco> enderecoOpt = enderecoRepository.findById(id);
-		if (enderecoOpt.isEmpty()){
-			throw new NotFoundException("Endereço de id=[" + id + "] não encontrado");
-		}
-		return mapper.map(enderecoOpt.get(),EnderecoResponseDto.class);
+		Endereco endereco = checkIfEnderecoExists(id);
+
+		return mapper.map(endereco,EnderecoResponseDto.class);
 	}
 
 	@Override
@@ -57,14 +52,27 @@ public class EnderecoService implements ICRUDService<EnderecoRequestDto, Enderec
 		this.findById(id);
 		Endereco endereco = mapper.map(dto, Endereco.class);
 		endereco.setId(id);
-		Hospital hospital = hospitalService.findById(dto.getHospitalId());
-		endereco.setHospital(hospital);
 		return mapper.map(enderecoRepository.save(endereco),EnderecoResponseDto.class);
 	}
 
 	@Override
 	public void deleteById(Long id) {
-		this.findById(id);
+		Endereco endereco = checkIfEnderecoExists(id);
+
+		if (Objects.nonNull(endereco.getHospital())) {
+			throw new BadRequestException("Não é possível deletar um endereço que está relacionado a um hospital");
+		}
+
 		enderecoRepository.deleteById(id);
+	}
+
+	private Endereco checkIfEnderecoExists(Long id) {
+		Optional<Endereco> enderecoOpt = enderecoRepository.findById(id);
+		
+		if (enderecoOpt.isEmpty()){
+			throw new NotFoundException("Endereço de id=[" + id + "] não encontrado");
+		}
+
+		return enderecoOpt.get();
 	}
 }
